Modularize src/ into focused modules for readability
Split the monolithic main.py (2018 lines) and pfsense_api_enhanced.py
(926 lines) into 14 focused modules under src/ and src/tools/. No
behavioral changes — all 114 tests pass, CLI entry point preserved.

New structure: models.py (enums/dataclasses), helpers.py (query helpers),
client.py (API client), server.py (mcp singleton), middleware.py (auth),
and src/tools/ with 8 domain modules (system, firewall, aliases, nat,
logs, services, dhcp, utility). Backward-compatible shims in main.py
and pfsense_api_enhanced.py preserve all existing imports.
