+++ src/main.py
@@ -6,1935 +6,67 @@ Compatible with pfSense REST API v2 (jaredhendrickson13/pfsense-api)
 """
 
 import asyncio
-import logging
 import os
 import sys
-from datetime import datetime
-from pathlib import Path
-from typing import Dict, List, Optional, Union
 
-# Load environment variables from .env file
-from dotenv import load_dotenv
+from .server import VERSION, get_api_client, logger, mcp
 
-# Determine the project root and load .env
-project_root = Path(__file__).parent.parent
-env_path = project_root / ".env"
-load_dotenv(dotenv_path=env_path)
-
-# MCP imports
-from fastmcp import FastMCP  # noqa: E402
-
-# Enhanced pfSense API client
-try:
-    from .pfsense_api_enhanced import (  # noqa: E402
-        AuthMethod,
-        ControlParameters,
-        EnhancedPfSenseAPIClient,
-        PaginationOptions,
-        PfSenseVersion,
-        QueryFilter,
-        SortOptions,
-        create_default_sort,
-        create_interface_filter,
-        create_pagination,
-        create_port_filter,
-    )
-except ImportError:
-    from pfsense_api_enhanced import (
-        AuthMethod,
-        ControlParameters,
-        EnhancedPfSenseAPIClient,
-        PaginationOptions,
-        PfSenseVersion,
-        QueryFilter,
-        SortOptions,
-        create_default_sort,
-        create_interface_filter,
-        create_pagination,
-        create_port_filter,
-    )
-
-# Configure logging
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-)
-logger = logging.getLogger(__name__)
-
-# Version
-VERSION = "4.0.0"
-
-# Initialize FastMCP server
-mcp = FastMCP(
-    "pfSense Enhanced MCP Server",
+# Import tool modules — each registers tools via @mcp.tool() on import
+from .tools import (  # noqa: F401
+    aliases,
+    dhcp,
+    firewall,
+    logs,
+    nat,
+    services,
+    system,
+    utility,
 )
 
-# Global API client
-api_client: Optional[EnhancedPfSenseAPIClient] = None
-
-def get_api_client() -> EnhancedPfSenseAPIClient:
-    """Get or create enhanced API client"""
-    global api_client
-    if api_client is None:
-        # Determine version
-        pf_version = os.getenv("PFSENSE_VERSION", "CE_2_8_0")
-        if pf_version == "PLUS_24_11":
-            version = PfSenseVersion.PLUS_24_11
-        else:
-            version = PfSenseVersion.CE_2_8_0
-
-        # Determine auth method
-        auth_method_str = os.getenv("AUTH_METHOD", "api_key").lower()
-        if auth_method_str == "basic":
-            auth_method = AuthMethod.BASIC
-        elif auth_method_str == "jwt":
-            auth_method = AuthMethod.JWT
-        else:
-            auth_method = AuthMethod.API_KEY
-
-        api_client = EnhancedPfSenseAPIClient(
-            host=os.getenv("PFSENSE_URL", "https://pfsense.local"),
-            auth_method=auth_method,
-            username=os.getenv("PFSENSE_USERNAME"),
-            password=os.getenv("PFSENSE_PASSWORD"),
-            api_key=os.getenv("PFSENSE_API_KEY"),
-            verify_ssl=os.getenv("VERIFY_SSL", "true").lower() == "true",
-            version=version,
-            enable_hateoas=os.getenv("ENABLE_HATEOAS", "false").lower() == "true"
-        )
-        logger.info(f"Enhanced API client initialized for pfSense {version.value}")
-    return api_client
-
-# Enhanced System Tools
-
-@mcp.tool()
-async def system_status() -> Dict:
-    """Get current system status including CPU, memory, disk usage, and version info"""
-    client = get_api_client()
-    try:
-        status = await client.get_system_status()
-
-        # Extract HATEOAS links if available
-        links = client.extract_links(status)
-
-        return {
-            "success": True,
-            "data": status.get("data", status),
-            "links": links,
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to get system status: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def search_interfaces(
-    search_term: Optional[str] = None,
-    status_filter: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "name"
-) -> Dict:
-    """Search and filter network interfaces with advanced options
-
-    Args:
-        search_term: Search in interface names/descriptions
-        status_filter: Filter by status (up, down, etc.)
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (name, status, etc.)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if search_term:
-            filters.append(QueryFilter("name", search_term, "contains"))
-
-        if status_filter:
-            filters.append(QueryFilter("status", status_filter))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by)
-
-        interfaces = await client.get_interfaces(
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "total_results": len(interfaces.get("data", [])),
-            "interfaces": interfaces.get("data", []),
-            "links": client.extract_links(interfaces),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search interfaces: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def find_interfaces_by_status(status: str) -> Dict:
-    """Find interfaces by their current status
-
-    Args:
-        status: Interface status to filter by (up, down, etc.)
-    """
-    client = get_api_client()
-    try:
-        interfaces = await client.find_interfaces_by_status(status)
-
-        return {
-            "success": True,
-            "status_filter": status,
-            "count": len(interfaces.get("data", [])),
-            "interfaces": interfaces.get("data", []),
-            "links": client.extract_links(interfaces),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to find interfaces by status: {e}")
-        return {"success": False, "error": str(e)}
-
-# Enhanced Firewall Tools
-
-@mcp.tool()
-async def search_firewall_rules(
-    interface: Optional[str] = None,
-    source_ip: Optional[str] = None,
-    destination_port: Optional[Union[int, str]] = None,
-    rule_type: Optional[str] = None,
-    search_description: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "tracker"
-) -> Dict:
-    """Search firewall rules with advanced filtering and pagination
-
-    Args:
-        interface: Filter by interface (wan, lan, etc.)
-        source_ip: Filter by source IP (supports partial matching)
-        destination_port: Filter by destination port
-        rule_type: Filter by rule type (pass, block, reject)
-        search_description: Search in rule descriptions
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (tracker, interface, type, descr, etc.)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if interface:
-            filters.append(create_interface_filter(interface))
-
-        if source_ip:
-            filters.append(QueryFilter("source", source_ip, "contains"))
-
-        if destination_port:
-            filters.append(create_port_filter(destination_port))
-
-        if rule_type:
-            filters.append(QueryFilter("type", rule_type))
-
-        if search_description:
-            filters.append(QueryFilter("descr", search_description, "contains"))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by)
-
-        rules = await client.get_firewall_rules(
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "interface": interface,
-                "source_ip": source_ip,
-                "destination_port": destination_port,
-                "rule_type": rule_type,
-                "search_description": search_description
-            },
-            "count": len(rules.get("data", [])),
-            "rules": rules.get("data", []),
-            "links": client.extract_links(rules),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search firewall rules: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def find_blocked_rules(
-    interface: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20
-) -> Dict:
-    """Find all firewall rules that block or reject traffic
-
-    Args:
-        interface: Optional interface filter
-        page: Page number for pagination
-        page_size: Number of results per page
-    """
-    client = get_api_client()
-    try:
-        create_pagination(page, page_size)
-        create_default_sort("tracker")
-
-        rules = await client.find_blocked_rules()
-
-        # Apply interface filter if specified
-        if interface:
-            filtered_rules = []
-            for rule in rules.get("data", []):
-                if rule.get("interface") == interface:
-                    filtered_rules.append(rule)
-            rules["data"] = filtered_rules
-
-        return {
-            "success": True,
-            "interface_filter": interface,
-            "count": len(rules.get("data", [])),
-            "blocked_rules": rules.get("data", []),
-            "links": client.extract_links(rules),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to find blocked rules: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def create_firewall_rule_advanced(
-    interface: str,
-    rule_type: str,
-    protocol: str,
-    source: str,
-    destination: str,
-    description: Optional[str] = None,
-    destination_port: Optional[str] = None,
-    position: Optional[int] = None,
-    apply_immediately: bool = True,
-    log_matches: bool = True
-) -> Dict:
-    """Create a firewall rule with advanced placement and control options
-
-    Args:
-        interface: Interface for the rule (wan, lan, etc.)
-        rule_type: Rule type (pass, block, reject)
-        protocol: Protocol (tcp, udp, icmp, any)
-        source: Source address (any, IP, network, alias)
-        destination: Destination address (any, IP, network, alias)
-        description: Optional rule description
-        destination_port: Optional destination port or range
-        position: Optional position to insert rule (0 = top)
-        apply_immediately: Whether to apply changes immediately
-        log_matches: Whether to log rule matches
-    """
-    client = get_api_client()
-
-    rule_data = {
-        "interface": [interface] if isinstance(interface, str) else interface,
-        "type": rule_type,
-        "ipprotocol": "inet",
-        "source": source,
-        "destination": destination,
-        "descr": description or f"Created via Enhanced MCP at {datetime.utcnow().isoformat()}",
-        "log": log_matches
-    }
-
-    # Handle protocol field - try null for "any", otherwise use the specified protocol
-    if protocol and protocol.lower() == "any":
-        rule_data["protocol"] = None  # Try null for "any protocol"
-    elif protocol:
-        rule_data["protocol"] = protocol
-
-    if destination_port:
-        rule_data["destination_port"] = destination_port
-
-    # Set control parameters
-    control = ControlParameters(
-        apply=apply_immediately,
-        placement=position
-    )
-
-    try:
-        result = await client.create_firewall_rule(rule_data, control)
-
-        return {
-            "success": True,
-            "message": "Firewall rule created with advanced options",
-            "rule": result.get("data", result),
-            "applied_immediately": apply_immediately,
-            "position": position,
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to create advanced firewall rule: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def move_firewall_rule(
-    rule_id: int,
-    new_position: int,
-    apply_immediately: bool = True
-) -> Dict:
-    """Move a firewall rule to a new position in the rule order
-
-    Args:
-        rule_id: Rule ID (array index from search_firewall_rules, e.g., 0, 1, 2...)
-        new_position: New position (0 = top, higher numbers = lower priority)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        result = await client.move_firewall_rule(
-            rule_id, new_position, apply_immediately
-        )
-
-        return {
-            "success": True,
-            "message": f"Rule {rule_id} moved to position {new_position}",
-            "rule_id": rule_id,
-            "new_position": new_position,
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to move firewall rule: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def update_firewall_rule(
-    rule_id: int,
-    rule_type: Optional[str] = None,
-    interface: Optional[str] = None,
-    protocol: Optional[str] = None,
-    source: Optional[str] = None,
-    destination: Optional[str] = None,
-    source_port: Optional[str] = None,
-    destination_port: Optional[str] = None,
-    description: Optional[str] = None,
-    disabled: Optional[bool] = None,
-    log_matches: Optional[bool] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Update an existing firewall rule by ID
-
-    Args:
-        rule_id: Rule ID (array index from search_firewall_rules, e.g., 0, 1, 2...)
-        rule_type: Rule type (pass, block, reject)
-        interface: Interface for the rule (wan, lan, etc.)
-        protocol: Protocol (tcp, udp, icmp, any)
-        source: Source address (any, IP, network, alias)
-        destination: Destination address (any, IP, network, alias)
-        source_port: Source port or range
-        destination_port: Destination port or range
-        description: Rule description
-        disabled: Whether the rule is disabled
-        log_matches: Whether to log rule matches
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        # Map Python parameter names to pfSense API field names
-        field_map = {
-            "rule_type": "type",
-            "interface": "interface",
-            "protocol": "protocol",
-            "source": "source",
-            "destination": "destination",
-            "source_port": "source_port",
-            "destination_port": "destination_port",
-            "description": "descr",
-            "disabled": "disabled",
-            "log_matches": "log",
-        }
-
-        params = {
-            "rule_type": rule_type,
-            "interface": interface,
-            "protocol": protocol,
-            "source": source,
-            "destination": destination,
-            "source_port": source_port,
-            "destination_port": destination_port,
-            "description": description,
-            "disabled": disabled,
-            "log_matches": log_matches,
-        }
-
-        updates = {}
-        for param_name, value in params.items():
-            if value is not None:
-                api_field = field_map[param_name]
-                if param_name == "interface":
-                    updates[api_field] = [value] if isinstance(value, str) else value
-                else:
-                    updates[api_field] = value
-
-        if not updates:
-            return {"success": False, "error": "No fields to update - provide at least one field"}
-
-        control = ControlParameters(apply=apply_immediately)
-        result = await client.update_firewall_rule(rule_id, updates, control)
-
-        return {
-            "success": True,
-            "message": f"Firewall rule {rule_id} updated",
-            "rule_id": rule_id,
-            "fields_updated": list(updates.keys()),
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to update firewall rule: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def delete_firewall_rule(
-    rule_id: int,
-    apply_immediately: bool = True
-) -> Dict:
-    """Delete a firewall rule by ID
-
-    Args:
-        rule_id: Rule ID (array index from search_firewall_rules, e.g., 0, 1, 2...)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        result = await client.delete_firewall_rule(rule_id, apply_immediately)
-
-        return {
-            "success": True,
-            "message": f"Firewall rule {rule_id} deleted",
-            "rule_id": rule_id,
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to delete firewall rule: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def bulk_block_ips(
-    ip_addresses: List[str],
-    interface: str = "wan",
-    description_prefix: str = "Bulk block via MCP"
-) -> Dict:
-    """Block multiple IP addresses at once
-
-    Args:
-        ip_addresses: List of IP addresses to block
-        interface: Interface to apply blocks on
-        description_prefix: Prefix for rule descriptions
-    """
-    client = get_api_client()
-    results = []
-    errors = []
-
-    for ip in ip_addresses:
-        try:
-            rule_data = {
-                "interface": [interface] if isinstance(interface, str) else interface,
-                "type": "block",
-                "ipprotocol": "inet",
-                "protocol": None,  # null = any protocol
-                "source": ip,
-                "destination": "any",
-                "descr": f"{description_prefix}: {ip}",
-                "log": True
-            }
-
-            # Don't apply immediately for bulk operations
-            control = ControlParameters(apply=False)
-            result = await client.create_firewall_rule(rule_data, control)
-            results.append({"ip": ip, "success": True, "rule_id": result.get("data", {}).get("id")})
-
-        except Exception as e:
-            logger.error(f"Failed to block IP {ip}: {e}")
-            errors.append({"ip": ip, "error": str(e)})
-
-    # Apply all changes at once
-    if results:
-        try:
-            await client._make_request("POST", "/firewall/apply")
-            applied = True
-        except Exception as e:
-            applied = False
-            logger.error(f"Failed to apply bulk changes: {e}")
-    else:
-        applied = False
-
-    return {
-        "success": len(results) > 0,
-        "total_requested": len(ip_addresses),
-        "successful": len(results),
-        "failed": len(errors),
-        "applied": applied,
-        "results": results,
-        "errors": errors,
-        "timestamp": datetime.utcnow().isoformat()
-    }
-
-# Enhanced Alias Tools
-
-@mcp.tool()
-async def search_aliases(
-    search_term: Optional[str] = None,
-    alias_type: Optional[str] = None,
-    containing_ip: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "name"
-) -> Dict:
-    """Search aliases with advanced filtering options
-
-    Args:
-        search_term: Search in alias names or descriptions
-        alias_type: Filter by alias type (host, network, port, url)
-        containing_ip: Find aliases containing this IP address
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if search_term:
-            filters.append(QueryFilter("name", search_term, "contains"))
-
-        if alias_type:
-            filters.append(QueryFilter("type", alias_type))
-
-        if containing_ip:
-            filters.append(QueryFilter("address", containing_ip, "contains"))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by)
-
-        aliases = await client.get_aliases(
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "search_term": search_term,
-                "alias_type": alias_type,
-                "containing_ip": containing_ip
-            },
-            "count": len(aliases.get("data", [])),
-            "aliases": aliases.get("data", []),
-            "links": client.extract_links(aliases),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search aliases: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def manage_alias_addresses(
-    alias_id: int,
-    action: str,
-    addresses: List[str]
-) -> Dict:
-    """Add or remove addresses from an existing alias
-
-    Args:
-        alias_id: ID of the alias to modify
-        action: Action to perform ('add' or 'remove')
-        addresses: List of addresses to add or remove
-    """
-    client = get_api_client()
-    try:
-        if action.lower() == "add":
-            result = await client.add_to_alias(alias_id, addresses)
-            message = f"Added {len(addresses)} addresses to alias {alias_id}"
-        elif action.lower() == "remove":
-            result = await client.remove_from_alias(alias_id, addresses)
-            message = f"Removed {len(addresses)} addresses from alias {alias_id}"
-        else:
-            return {"success": False, "error": "Action must be 'add' or 'remove'"}
-
-        return {
-            "success": True,
-            "message": message,
-            "alias_id": alias_id,
-            "action": action,
-            "addresses": addresses,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to manage alias addresses: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def create_alias(
-    name: str,
-    alias_type: str,
-    addresses: List[str],
-    description: Optional[str] = None,
-    details: Optional[List[str]] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Create a new firewall alias
-
-    Args:
-        name: Alias name (max 31 characters, alphanumeric and underscores)
-        alias_type: Type of alias (host, network, port)
-        addresses: List of initial entries (IPs, networks, or ports depending on type)
-        description: Optional description of the alias
-        details: Optional per-entry descriptions (same length as addresses)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        alias_data = {
-            "name": name,
-            "type": alias_type,
-            "address": addresses,
-        }
-
-        if description:
-            alias_data["descr"] = description
-
-        if details:
-            alias_data["detail"] = details
-
-        control = ControlParameters(apply=apply_immediately)
-        result = await client.create_alias(alias_data, control)
-
-        return {
-            "success": True,
-            "message": f"Alias '{name}' created",
-            "alias": result.get("data", result),
-            "applied": apply_immediately,
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to create alias: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def update_alias(
-    alias_id: int,
-    name: Optional[str] = None,
-    alias_type: Optional[str] = None,
-    addresses: Optional[List[str]] = None,
-    description: Optional[str] = None,
-    details: Optional[List[str]] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Update an existing alias by ID
-
-    Args:
-        alias_id: Alias ID (array index from search_aliases)
-        name: New alias name
-        alias_type: New alias type (host, network, port)
-        addresses: New list of addresses
-        description: New description
-        details: New per-entry descriptions
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        field_map = {
-            "name": "name",
-            "alias_type": "type",
-            "addresses": "address",
-            "description": "descr",
-            "details": "detail",
-        }
-
-        params = {
-            "name": name,
-            "alias_type": alias_type,
-            "addresses": addresses,
-            "description": description,
-            "details": details,
-        }
-
-        updates = {}
-        for param_name, value in params.items():
-            if value is not None:
-                updates[field_map[param_name]] = value
-
-        if not updates:
-            return {"success": False, "error": "No fields to update - provide at least one field"}
-
-        control = ControlParameters(apply=apply_immediately)
-        result = await client.update_alias(alias_id, updates, control)
-
-        return {
-            "success": True,
-            "message": f"Alias {alias_id} updated",
-            "alias_id": alias_id,
-            "fields_updated": list(updates.keys()),
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to update alias: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def delete_alias(
-    alias_id: int,
-    apply_immediately: bool = True
-) -> Dict:
-    """Delete an alias by ID
-
-    Args:
-        alias_id: Alias ID (array index from search_aliases)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        result = await client.delete_alias(alias_id, apply_immediately)
-
-        return {
-            "success": True,
-            "message": f"Alias {alias_id} deleted",
-            "alias_id": alias_id,
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to delete alias: {e}")
-        return {"success": False, "error": str(e)}
-
-# NAT Port Forward Tools
-
-@mcp.tool()
-async def search_nat_port_forwards(
-    interface: Optional[str] = None,
-    protocol: Optional[str] = None,
-    destination_port: Optional[Union[int, str]] = None,
-    search_description: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "interface"
-) -> Dict:
-    """Search NAT port forwarding rules with filtering and pagination
-
-    Args:
-        interface: Filter by interface (wan, lan, etc.)
-        protocol: Filter by protocol (tcp, udp, tcp/udp)
-        destination_port: Filter by external destination port
-        search_description: Search in rule descriptions
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (interface, protocol, descr, etc.)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if interface:
-            filters.append(create_interface_filter(interface))
-
-        if protocol:
-            filters.append(QueryFilter("protocol", protocol))
-
-        if destination_port:
-            filters.append(QueryFilter("destination_port", str(destination_port)))
-
-        if search_description:
-            filters.append(QueryFilter("descr", search_description, "contains"))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by)
-
-        result = await client._make_request(
-            "GET", "/firewall/nat/port_forwards",
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "interface": interface,
-                "protocol": protocol,
-                "destination_port": destination_port,
-                "search_description": search_description
-            },
-            "count": len(result.get("data", [])),
-            "port_forwards": result.get("data", []),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search NAT port forwards: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def create_nat_port_forward(
-    interface: str,
-    protocol: str,
-    destination: str,
-    destination_port: str,
-    target: str,
-    local_port: str,
-    source: str = "any",
-    description: Optional[str] = None,
-    disabled: bool = False,
-    nat_reflection: Optional[str] = None,
-    create_associated_rule: bool = True,
-    apply_immediately: bool = True
-) -> Dict:
-    """Create a NAT port forwarding rule
-
-    Args:
-        interface: Interface for the rule (typically "wan")
-        protocol: Protocol (tcp, udp, tcp/udp)
-        destination: External address (e.g., "wanip", "any", or specific IP/alias)
-        destination_port: External port or range (e.g., "80", "8080-8090")
-        target: Internal host IP address or alias to forward to
-        local_port: Internal port to forward to (e.g., "80")
-        source: Source address filter (default: "any")
-        description: Optional rule description
-        disabled: Whether the rule starts disabled
-        nat_reflection: NAT reflection mode (enable, disable, purenat)
-        create_associated_rule: Whether to create an associated firewall pass rule
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        forward_data = {
-            "interface": [interface] if isinstance(interface, str) else interface,
-            "protocol": protocol,
-            "destination": destination,
-            "destination_port": destination_port,
-            "target": target,
-            "local_port": local_port,
-            "source": source,
-            "disabled": disabled,
-            "associated_rule_id": "new" if create_associated_rule else "",
-        }
-
-        if description:
-            forward_data["descr"] = description
-        else:
-            forward_data["descr"] = f"Port forward via MCP at {datetime.utcnow().isoformat()}"
-
-        if nat_reflection:
-            forward_data["natreflection"] = nat_reflection
-
-        control = ControlParameters(apply=apply_immediately)
-
-        result = await client._make_request(
-            "POST", "/firewall/nat/port_forward",
-            data=forward_data, control=control
-        )
-
-        return {
-            "success": True,
-            "message": f"NAT port forward created: {interface} {protocol}/{destination_port} -> {target}:{local_port}",
-            "port_forward": result.get("data", result),
-            "applied": apply_immediately,
-            "associated_rule_created": create_associated_rule,
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to create NAT port forward: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def delete_nat_port_forward(
-    port_forward_id: int,
-    apply_immediately: bool = True
-) -> Dict:
-    """Delete a NAT port forwarding rule by ID
-
-    Args:
-        port_forward_id: Port forward rule ID (array index from search_nat_port_forwards)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        control = ControlParameters(apply=apply_immediately)
-
-        result = await client._make_request(
-            "DELETE", "/firewall/nat/port_forward",
-            data={"id": port_forward_id}, control=control
-        )
-
-        return {
-            "success": True,
-            "message": f"NAT port forward {port_forward_id} deleted",
-            "port_forward_id": port_forward_id,
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to delete NAT port forward: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def update_nat_port_forward(
-    port_forward_id: int,
-    interface: Optional[str] = None,
-    protocol: Optional[str] = None,
-    destination: Optional[str] = None,
-    destination_port: Optional[str] = None,
-    target: Optional[str] = None,
-    local_port: Optional[str] = None,
-    source: Optional[str] = None,
-    description: Optional[str] = None,
-    disabled: Optional[bool] = None,
-    nat_reflection: Optional[str] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Update an existing NAT port forwarding rule by ID
-
-    Args:
-        port_forward_id: Port forward rule ID (array index from search_nat_port_forwards)
-        interface: Interface for the rule (wan, lan, etc.)
-        protocol: Protocol (tcp, udp, tcp/udp)
-        destination: External address (e.g., "wanip", "any", or specific IP/alias)
-        destination_port: External port or range
-        target: Internal host IP address or alias
-        local_port: Internal port to forward to
-        source: Source address filter
-        description: Rule description
-        disabled: Whether the rule is disabled
-        nat_reflection: NAT reflection mode (enable, disable, purenat)
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        field_map = {
-            "interface": "interface",
-            "protocol": "protocol",
-            "destination": "destination",
-            "destination_port": "destination_port",
-            "target": "target",
-            "local_port": "local_port",
-            "source": "source",
-            "description": "descr",
-            "disabled": "disabled",
-            "nat_reflection": "natreflection",
-        }
-
-        params = {
-            "interface": interface,
-            "protocol": protocol,
-            "destination": destination,
-            "destination_port": destination_port,
-            "target": target,
-            "local_port": local_port,
-            "source": source,
-            "description": description,
-            "disabled": disabled,
-            "nat_reflection": nat_reflection,
-        }
-
-        updates = {}
-        for param_name, value in params.items():
-            if value is not None:
-                api_field = field_map[param_name]
-                if param_name == "interface":
-                    updates[api_field] = [value] if isinstance(value, str) else value
-                else:
-                    updates[api_field] = value
-
-        if not updates:
-            return {"success": False, "error": "No fields to update - provide at least one field"}
-
-        updates["id"] = port_forward_id
-        control = ControlParameters(apply=apply_immediately)
-
-        result = await client._make_request(
-            "PATCH", "/firewall/nat/port_forward",
-            data=updates, control=control
-        )
-
-        return {
-            "success": True,
-            "message": f"NAT port forward {port_forward_id} updated",
-            "port_forward_id": port_forward_id,
-            "fields_updated": [k for k in updates.keys() if k != "id"],
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to update NAT port forward: {e}")
-        return {"success": False, "error": str(e)}
-
-# Enhanced Log Analysis Tools
-
-@mcp.tool()
-async def get_firewall_log(
-    lines: int = 20,
-    action_filter: Optional[str] = None,
-    interface: Optional[str] = None,
-    source_ip: Optional[str] = None,
-    destination_ip: Optional[str] = None,
-    destination_port: Optional[str] = None,
-    protocol: Optional[str] = None,
-    sort_by: str = "timestamp"
-) -> Dict:
-    """Get firewall log entries with optional filtering
-
-    Args:
-        lines: Number of log lines to retrieve (default 20, max 50)
-        action_filter: Filter by action (pass, block, reject)
-        interface: Filter by interface (wan, lan, etc.)
-        source_ip: Filter by source IP address
-        destination_ip: Filter by destination IP address
-        destination_port: Filter by destination port
-        protocol: Filter by protocol (tcp, udp, icmp)
-        sort_by: Field to sort by (default "timestamp" descending)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if action_filter:
-            filters.append(QueryFilter("action", action_filter))
-        if interface:
-            filters.append(QueryFilter("interface", interface))
-        if source_ip:
-            filters.append(QueryFilter("src_ip", source_ip))
-        if destination_ip:
-            filters.append(QueryFilter("dst_ip", destination_ip))
-        if destination_port:
-            filters.append(QueryFilter("dst_port", destination_port))
-        if protocol:
-            filters.append(QueryFilter("protocol", protocol))
-
-        sort = create_default_sort(sort_by, descending=True)
-
-        logs = await client.get_firewall_logs(
-            lines=min(lines, 50),
-            filters=filters if filters else None,
-            sort=sort,
-        )
-
-        return {
-            "success": True,
-            "lines_requested": min(lines, 50),
-            "filters_applied": {
-                "action": action_filter,
-                "interface": interface,
-                "source_ip": source_ip,
-                "destination_ip": destination_ip,
-                "destination_port": destination_port,
-                "protocol": protocol,
-            },
-            "count": len(logs.get("data", [])),
-            "log_entries": logs.get("data", []),
-            "links": client.extract_links(logs),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to get firewall log: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def analyze_blocked_traffic(
-    hours_back: int = 24,
-    limit: int = 20,
-    group_by_source: bool = True
-) -> Dict:
-    """Analyze blocked traffic patterns from firewall logs
-
-    Args:
-        hours_back: How many hours back to analyze
-        limit: Maximum number of log entries to analyze
-        group_by_source: Whether to group results by source IP
-    """
-    client = get_api_client()
-    try:
-        # Get blocked traffic logs
-        create_default_sort("timestamp", descending=True)
-        logs = await client.get_blocked_traffic_logs(lines=limit)
-
-        log_data = logs.get("data", [])
-
-        if group_by_source:
-            # Group by source IP
-            source_stats = {}
-            for entry in log_data:
-                src_ip = entry.get("src_ip", "unknown")
-                if src_ip not in source_stats:
-                    source_stats[src_ip] = {
-                        "count": 0,
-                        "ports": set(),
-                        "destinations": set(),
-                        "latest_time": None
-                    }
-
-                source_stats[src_ip]["count"] += 1
-
-                if entry.get("dst_port"):
-                    source_stats[src_ip]["ports"].add(entry["dst_port"])
-
-                if entry.get("dst_ip"):
-                    source_stats[src_ip]["destinations"].add(entry["dst_ip"])
-
-                if entry.get("timestamp"):
-                    source_stats[src_ip]["latest_time"] = entry["timestamp"]
-
-            # Convert sets to lists for JSON serialization
-            for ip, stats in source_stats.items():
-                stats["ports"] = list(stats["ports"])
-                stats["destinations"] = list(stats["destinations"])
-                stats["threat_score"] = min(10, stats["count"] / 10)
-
-            # Sort by count
-            sorted_sources = sorted(
-                source_stats.items(),
-                key=lambda x: x[1]["count"],
-                reverse=True
-            )
-
-            analysis = {
-                "grouped_by": "source_ip",
-                "total_unique_sources": len(source_stats),
-                "top_sources": dict(sorted_sources[:20])
-            }
-        else:
-            analysis = {
-                "grouped_by": "none",
-                "raw_entries": log_data
-            }
-
-        return {
-            "success": True,
-            "analysis_period_hours": hours_back,
-            "total_entries_analyzed": len(log_data),
-            "analysis": analysis,
-            "links": client.extract_links(logs),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to analyze blocked traffic: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def search_logs_by_ip(
-    ip_address: str,
-    log_type: str = "firewall",
-    lines: int = 100,
-    include_related: bool = True
-) -> Dict:
-    """Search logs for activity related to a specific IP address
-
-    Args:
-        ip_address: IP address to search for
-        log_type: Type of logs to search (firewall, system, etc.)
-        lines: Number of log lines to retrieve
-        include_related: Whether to include related traffic (src and dst)
-    """
-    client = get_api_client()
-    try:
-        if log_type == "firewall":
-            logs = await client.get_logs_by_ip(ip_address, lines)
-        else:
-            # For other log types, use general log search
-            filters = [QueryFilter("message", ip_address, "contains")]
-            logs = await client._make_request(
-                "GET", f"/diagnostics/log/{log_type}",
-                filters=filters,
-                pagination=PaginationOptions(limit=lines)
-            )
-
-        log_entries = logs.get("data", [])
-
-        # Analyze patterns if firewall logs
-        if log_type == "firewall" and log_entries:
-            patterns = {
-                "total_entries": len(log_entries),
-                "blocked_count": 0,
-                "allowed_count": 0,
-                "ports_accessed": set(),
-                "protocols_used": set()
-            }
-
-            for entry in log_entries:
-                action = entry.get("action", "").lower()
-                if "block" in action or "reject" in action:
-                    patterns["blocked_count"] += 1
-                elif "pass" in action or "allow" in action:
-                    patterns["allowed_count"] += 1
-
-                if entry.get("dst_port"):
-                    patterns["ports_accessed"].add(entry["dst_port"])
-
-                if entry.get("protocol"):
-                    patterns["protocols_used"].add(entry["protocol"])
-
-            # Convert sets to lists
-            patterns["ports_accessed"] = list(patterns["ports_accessed"])
-            patterns["protocols_used"] = list(patterns["protocols_used"])
-        else:
-            patterns = None
-
-        return {
-            "success": True,
-            "ip_address": ip_address,
-            "log_type": log_type,
-            "total_entries": len(log_entries),
-            "patterns": patterns,
-            "log_entries": log_entries,
-            "links": client.extract_links(logs),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search logs by IP: {e}")
-        return {"success": False, "error": str(e)}
-
-# Service Tools
-
-@mcp.tool()
-async def search_services(
-    search_term: Optional[str] = None,
-    status_filter: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "name"
-) -> Dict:
-    """Search and filter system services
-
-    Args:
-        search_term: Search in service names or descriptions
-        status_filter: Filter by status (running, stopped)
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (name, status, description)
-    """
-    client = get_api_client()
-    try:
-        if status_filter == "running":
-            result = await client.find_running_services()
-        elif status_filter == "stopped":
-            result = await client.find_stopped_services()
-        else:
-            result = await client.get_services()
-
-        services = result.get("data", [])
-
-        if search_term:
-            term_lower = search_term.lower()
-            services = [
-                s for s in services
-                if term_lower in s.get("name", "").lower()
-                or term_lower in s.get("description", "").lower()
-            ]
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "search_term": search_term,
-                "status_filter": status_filter,
-            },
-            "count": len(services),
-            "services": services,
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search services: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def control_service(
-    service_name: str,
-    action: str
-) -> Dict:
-    """Start, stop, or restart a system service
-
-    Args:
-        service_name: Name of the service (e.g., "dhcpd", "unbound", "ntpd")
-        action: Action to perform ("start", "stop", or "restart")
-    """
-    client = get_api_client()
-    try:
-        action_lower = action.lower()
-        if action_lower == "start":
-            result = await client.start_service(service_name)
-        elif action_lower == "stop":
-            result = await client.stop_service(service_name)
-        elif action_lower == "restart":
-            result = await client.restart_service(service_name)
-        else:
-            return {
-                "success": False,
-                "error": f"Invalid action '{action}'. Must be 'start', 'stop', or 'restart'",
-            }
-
-        return {
-            "success": True,
-            "message": f"Service '{service_name}' {action_lower} command sent",
-            "service": service_name,
-            "action": action_lower,
-            "result": result.get("data", result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to {action} service {service_name}: {e}")
-        return {"success": False, "error": str(e)}
-
-# Enhanced DHCP Tools
-
-@mcp.tool()
-async def search_dhcp_leases(
-    search_term: Optional[str] = None,
-    interface: Optional[str] = None,
-    mac_address: Optional[str] = None,
-    hostname: Optional[str] = None,
-    state: str = "active",
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "starts"
-) -> Dict:
-    """Search DHCP leases with advanced filtering
-
-    Args:
-        search_term: General search term for hostname or IP
-        interface: Filter by interface
-        mac_address: Filter by specific MAC address
-        hostname: Filter by hostname (supports partial matching)
-        state: Filter by lease state (active, expired, etc.)
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (starts, ends, hostname, ip, mac)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if search_term:
-            filters.append(QueryFilter("hostname", search_term, "contains"))
-
-        if interface:
-            # DHCP uses 'if' field, not 'interface'
-            filters.append(QueryFilter("if", interface, "contains"))
-
-        if mac_address:
-            filters.append(QueryFilter("mac", mac_address))
-
-        if hostname:
-            filters.append(QueryFilter("hostname", hostname, "contains"))
-
-        if state:
-            # DHCP uses 'active_status' field, not 'state'
-            filters.append(QueryFilter("active_status", state))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by, descending=True)
-
-        leases = await client.get_dhcp_leases(
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "search_term": search_term,
-                "interface": interface,
-                "mac_address": mac_address,
-                "hostname": hostname,
-                "state": state
-            },
-            "count": len(leases.get("data", [])),
-            "leases": leases.get("data", []),
-            "links": client.extract_links(leases),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search DHCP leases: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def search_dhcp_static_mappings(
-    interface: Optional[str] = None,
-    mac_address: Optional[str] = None,
-    hostname: Optional[str] = None,
-    ip_address: Optional[str] = None,
-    page: int = 1,
-    page_size: int = 20,
-    sort_by: str = "mac"
-) -> Dict:
-    """Search DHCP static mappings (reservations) with filtering
-
-    Args:
-        interface: Filter by interface (lan, opt1, etc.)
-        mac_address: Filter by MAC address
-        hostname: Filter by hostname (partial match)
-        ip_address: Filter by IP address
-        page: Page number for pagination
-        page_size: Number of results per page
-        sort_by: Field to sort by (mac, ipaddr, hostname)
-    """
-    client = get_api_client()
-    try:
-        filters = []
-
-        if interface:
-            filters.append(QueryFilter("parent_id", interface))
-
-        if mac_address:
-            filters.append(QueryFilter("mac", mac_address))
-
-        if hostname:
-            filters.append(QueryFilter("hostname", hostname, "contains"))
-
-        if ip_address:
-            filters.append(QueryFilter("ipaddr", ip_address))
-
-        pagination = create_pagination(page, page_size)
-        sort = create_default_sort(sort_by)
-
-        result = await client.get_dhcp_static_mappings(
-            filters=filters if filters else None,
-            sort=sort,
-            pagination=pagination
-        )
-
-        return {
-            "success": True,
-            "page": page,
-            "page_size": page_size,
-            "filters_applied": {
-                "interface": interface,
-                "mac_address": mac_address,
-                "hostname": hostname,
-                "ip_address": ip_address,
-            },
-            "count": len(result.get("data", [])),
-            "static_mappings": result.get("data", []),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to search DHCP static mappings: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def create_dhcp_static_mapping(
-    interface: str,
-    mac_address: str,
-    ip_address: str,
-    hostname: Optional[str] = None,
-    description: Optional[str] = None,
-    domain: Optional[str] = None,
-    gateway: Optional[str] = None,
-    dns_server: Optional[str] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Create a DHCP static mapping (reservation)
-
-    Args:
-        interface: Interface/DHCP pool (e.g., "lan")
-        mac_address: MAC address to reserve for
-        ip_address: IP address to assign
-        hostname: Optional hostname
-        description: Optional description
-        domain: Optional domain name
-        gateway: Optional gateway override
-        dns_server: Optional DNS server override
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        mapping_data = {
-            "parent_id": interface,
-            "mac": mac_address,
-            "ipaddr": ip_address,
-        }
-
-        if hostname:
-            mapping_data["hostname"] = hostname
-        if description:
-            mapping_data["descr"] = description
-        if domain:
-            mapping_data["domain"] = domain
-        if gateway:
-            mapping_data["gateway"] = gateway
-        if dns_server:
-            mapping_data["dnsserver"] = dns_server
-
-        control = ControlParameters(apply=apply_immediately)
-        result = await client.create_dhcp_static_mapping(mapping_data, control)
-
-        return {
-            "success": True,
-            "message": f"DHCP static mapping created: {mac_address} -> {ip_address}",
-            "static_mapping": result.get("data", result),
-            "applied": apply_immediately,
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to create DHCP static mapping: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def update_dhcp_static_mapping(
-    mapping_id: int,
-    mac_address: Optional[str] = None,
-    ip_address: Optional[str] = None,
-    hostname: Optional[str] = None,
-    description: Optional[str] = None,
-    interface: Optional[str] = None,
-    apply_immediately: bool = True
-) -> Dict:
-    """Update an existing DHCP static mapping by ID
-
-    Args:
-        mapping_id: Static mapping ID
-        mac_address: New MAC address
-        ip_address: New IP address
-        hostname: New hostname
-        description: New description
-        interface: New interface/DHCP pool
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        field_map = {
-            "mac_address": "mac",
-            "ip_address": "ipaddr",
-            "hostname": "hostname",
-            "description": "descr",
-            "interface": "parent_id",
-        }
-
-        params = {
-            "mac_address": mac_address,
-            "ip_address": ip_address,
-            "hostname": hostname,
-            "description": description,
-            "interface": interface,
-        }
-
-        updates = {}
-        for param_name, value in params.items():
-            if value is not None:
-                updates[field_map[param_name]] = value
-
-        if not updates:
-            return {"success": False, "error": "No fields to update - provide at least one field"}
-
-        control = ControlParameters(apply=apply_immediately)
-        result = await client.update_dhcp_static_mapping(mapping_id, updates, control)
-
-        return {
-            "success": True,
-            "message": f"DHCP static mapping {mapping_id} updated",
-            "mapping_id": mapping_id,
-            "fields_updated": list(updates.keys()),
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to update DHCP static mapping: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def delete_dhcp_static_mapping(
-    mapping_id: int,
-    apply_immediately: bool = True
-) -> Dict:
-    """Delete a DHCP static mapping by ID
-
-    Args:
-        mapping_id: Static mapping ID
-        apply_immediately: Whether to apply changes immediately
-    """
-    client = get_api_client()
-    try:
-        result = await client.delete_dhcp_static_mapping(mapping_id, apply_immediately)
-
-        return {
-            "success": True,
-            "message": f"DHCP static mapping {mapping_id} deleted",
-            "mapping_id": mapping_id,
-            "applied": apply_immediately,
-            "result": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to delete DHCP static mapping: {e}")
-        return {"success": False, "error": str(e)}
-
-# HATEOAS Navigation Tools
-
-@mcp.tool()
-async def follow_api_link(link_url: str) -> Dict:
-    """Follow a HATEOAS link from a previous API response
-
-    Args:
-        link_url: The link URL to follow (from _links section)
-    """
-    client = get_api_client()
-    try:
-        result = await client.follow_link(link_url)
-
-        return {
-            "success": True,
-            "followed_link": link_url,
-            "data": result.get("data", result),
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to follow link: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def enable_hateoas() -> Dict:
-    """Enable HATEOAS links in API responses for this session"""
-    client = get_api_client()
-    result = await client.enable_hateoas()
-    return {
-        "success": True,
-        "message": "HATEOAS enabled - API responses will now include navigation links",
-        "result": result,
-        "timestamp": datetime.utcnow().isoformat()
-    }
-
-@mcp.tool()
-async def disable_hateoas() -> Dict:
-    """Disable HATEOAS links in API responses for this session"""
-    client = get_api_client()
-    result = await client.disable_hateoas()
-    return {
-        "success": True,
-        "message": "HATEOAS disabled - API responses will be more compact",
-        "result": result,
-        "timestamp": datetime.utcnow().isoformat()
-    }
-
-# Object ID Management Tools
-
-@mcp.tool()
-async def refresh_object_ids(endpoint: str) -> Dict:
-    """Refresh object IDs by re-querying an endpoint (handles ID changes after deletions)
-
-    Args:
-        endpoint: API endpoint to refresh (e.g., '/firewall/rule')
-    """
-    client = get_api_client()
-    try:
-        result = await client.refresh_object_ids(endpoint)
-
-        return {
-            "success": True,
-            "endpoint": endpoint,
-            "refreshed_count": len(result.get("data", [])),
-            "objects": result.get("data", []),
-            "message": "Object IDs refreshed - use updated IDs for future operations",
-            "links": client.extract_links(result),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to refresh object IDs: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def find_object_by_field(
-    endpoint: str,
-    field: str,
-    value: str
-) -> Dict:
-    """Find an object by a specific field value (safer than using IDs)
-
-    Args:
-        endpoint: API endpoint to search
-        field: Field name to search by
-        value: Value to search for
-    """
-    client = get_api_client()
-    try:
-        obj = await client.find_object_by_field(endpoint, field, value)
-
-        if obj:
-            return {
-                "success": True,
-                "endpoint": endpoint,
-                "search_field": field,
-                "search_value": value,
-                "found": True,
-                "object": obj,
-                "object_id": obj.get("id"),
-                "timestamp": datetime.utcnow().isoformat()
-            }
-        else:
-            return {
-                "success": True,
-                "endpoint": endpoint,
-                "search_field": field,
-                "search_value": value,
-                "found": False,
-                "message": "No object found matching criteria",
-                "timestamp": datetime.utcnow().isoformat()
-            }
-    except Exception as e:
-        logger.error(f"Failed to find object by field: {e}")
-        return {"success": False, "error": str(e)}
-
-# API Capabilities and Testing
-
-@mcp.tool()
-async def get_api_capabilities() -> Dict:
-    """Get comprehensive API capabilities and configuration"""
-    client = get_api_client()
-    try:
-        capabilities = await client.get_api_capabilities()
-
-        return {
-            "success": True,
-            "api_version": "v2",
-            "package": "jaredhendrickson13/pfsense-api",
-            "pfsense_version": os.getenv("PFSENSE_VERSION", "CE_2_8_0"),
-            "capabilities": capabilities.get("data", capabilities),
-            "features": {
-                "object_ids": "Dynamic, non-persistent",
-                "queries_filters": "Full support with multiple operators",
-                "sorting": "Multi-field sorting supported",
-                "pagination": "Limit/offset based",
-                "hateoas": f"{'Enabled' if client.hateoas_enabled else 'Disabled'}",
-                "control_parameters": "Apply, async, placement, append, remove"
-            },
-            "links": client.extract_links(capabilities),
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Failed to get API capabilities: {e}")
-        return {"success": False, "error": str(e)}
-
-@mcp.tool()
-async def test_enhanced_connection() -> Dict:
-    """Test enhanced API connection with feature validation"""
-    client = get_api_client()
-    try:
-        # Test basic connection
-        connected = await client.test_connection()
-
-        if not connected:
-            return {
-                "success": False,
-                "message": "Basic connection failed",
-                "timestamp": datetime.utcnow().isoformat()
-            }
-
-        # Test advanced features
-        tests = []
-
-        # Test filtering
-        try:
-            await client.get_interfaces(
-                filters=[QueryFilter("status", "up")],
-                pagination=PaginationOptions(limit=1)
-            )
-            tests.append({"feature": "filtering", "status": "working"})
-        except Exception as e:
-            tests.append({"feature": "filtering", "status": "failed", "error": str(e)})
-
-        # Test sorting
-        try:
-            await client.get_firewall_rules(
-                sort=SortOptions(sort_by="interface"),
-                pagination=PaginationOptions(limit=1)
-            )
-            tests.append({"feature": "sorting", "status": "working"})
-        except Exception as e:
-            tests.append({"feature": "sorting", "status": "failed", "error": str(e)})
-
-        # Test HATEOAS if enabled
-        if client.hateoas_enabled:
-            try:
-                result = await client.get_system_status()
-                links = client.extract_links(result)
-                if links:
-                    tests.append({"feature": "hateoas", "status": "working", "links_found": len(links)})
-                else:
-                    tests.append({"feature": "hateoas", "status": "no_links"})
-            except Exception as e:
-                tests.append({"feature": "hateoas", "status": "failed", "error": str(e)})
-
-        working_features = len([t for t in tests if t["status"] == "working"])
-
-        return {
-            "success": True,
-            "message": f"Enhanced connection test completed - {working_features}/{len(tests)} features working",
-            "basic_connection": True,
-            "feature_tests": tests,
-            "hateoas_enabled": client.hateoas_enabled,
-            "timestamp": datetime.utcnow().isoformat()
-        }
-    except Exception as e:
-        logger.error(f"Enhanced connection test failed: {e}")
-        return {"success": False, "error": str(e)}
-
-# Bearer token auth middleware for HTTP transport
-class BearerAuthMiddleware:
-    """ASGI middleware that checks for a valid Bearer token in the Authorization header."""
-
-    def __init__(self, app, api_key: str):
-        self.app = app
-        self.api_key = api_key
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "http":
-            headers = dict(scope.get("headers", []))
-            auth_header = headers.get(b"authorization", b"").decode()
-            if not auth_header.startswith("Bearer ") or auth_header[7:] != self.api_key:
-                from starlette.responses import JSONResponse
-                response = JSONResponse({"error": "Unauthorized"}, status_code=401)
-                await response(scope, receive, send)
-                return
-        await self.app(scope, receive, send)
+# Re-export all tool FunctionTool objects so ``_main.system_status`` still works in tests
+system_status = system.system_status
+search_interfaces = system.search_interfaces
+find_interfaces_by_status = system.find_interfaces_by_status
+
+search_firewall_rules = firewall.search_firewall_rules
+find_blocked_rules = firewall.find_blocked_rules
+create_firewall_rule_advanced = firewall.create_firewall_rule_advanced
+move_firewall_rule = firewall.move_firewall_rule
+update_firewall_rule = firewall.update_firewall_rule
+delete_firewall_rule = firewall.delete_firewall_rule
+bulk_block_ips = firewall.bulk_block_ips
+
+search_aliases = aliases.search_aliases
+manage_alias_addresses = aliases.manage_alias_addresses
+create_alias = aliases.create_alias
+update_alias = aliases.update_alias
+delete_alias = aliases.delete_alias
+
+search_nat_port_forwards = nat.search_nat_port_forwards
+create_nat_port_forward = nat.create_nat_port_forward
+delete_nat_port_forward = nat.delete_nat_port_forward
+update_nat_port_forward = nat.update_nat_port_forward
+
+get_firewall_log = logs.get_firewall_log
+analyze_blocked_traffic = logs.analyze_blocked_traffic
+search_logs_by_ip = logs.search_logs_by_ip
+
+search_services = services.search_services
+control_service = services.control_service
+
+search_dhcp_leases = dhcp.search_dhcp_leases
+search_dhcp_static_mappings = dhcp.search_dhcp_static_mappings
+create_dhcp_static_mapping = dhcp.create_dhcp_static_mapping
+update_dhcp_static_mapping = dhcp.update_dhcp_static_mapping
+delete_dhcp_static_mapping = dhcp.delete_dhcp_static_mapping
+
+follow_api_link = utility.follow_api_link
+enable_hateoas = utility.enable_hateoas
+disable_hateoas = utility.disable_hateoas
+refresh_object_ids = utility.refresh_object_ids
+find_object_by_field = utility.find_object_by_field
+get_api_capabilities = utility.get_api_capabilities
+test_enhanced_connection = utility.test_enhanced_connection
 
 
 # Main execution
@@ -1995,6 +127,8 @@ def main():
         mcp.run(transport="stdio")
     elif args.transport == "streamable-http":
         import uvicorn
+
+        from .middleware import BearerAuthMiddleware
 
         # Use sse_app() for FastMCP < 2.14, http_app() for >= 2.14
         if hasattr(mcp, 'http_app'):
