+++ src/tools/__init__.py
@@ -1,0 +1,1 @@
+"""pfSense MCP tool modules — each file registers tools via @mcp.tool()."""
